Add test for service layer

+++ src/main/java/backend/storage/api/service/impl/ProductServiceImpl.java
@@ -6,11 +6,10 @@ import backend.storage.api.exception.EntityNotFoundException;
 import backend.storage.api.mapper.ProductMapper;
 import backend.storage.api.model.Product;
 import backend.storage.api.repository.ProductRepository;
-import backend.storage.api.repository.ProductTypeRepository;
 import backend.storage.api.service.ProductService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -27,7 +26,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<ProductResponseDto> findAll() {
+    public List<ProductResponseDto> findAll(Pageable pageable) {
         return productRepository.findAll()
                 .stream()
                 .map(productMapper::toResponseFromEntity)
@@ -37,13 +36,13 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public ProductResponseDto findById(Long id) {
         return productMapper.toResponseFromEntity(productRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Product not found")));
+                .orElseThrow(() -> new EntityNotFoundException("Product with id: " + id + " not found")));
     }
 
     @Override
     public ProductResponseDto updateProductById(Long id, ProductRequestDto requestDto) {
-        Product currentProduct = productRepository.getReferenceById(id);
-        // .orElseThrow(() -> new EntityNotFoundException("Can't get product by id: " + id));
+        Product currentProduct = productRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Can't find product with id: " + id));
         currentProduct.setProductType(requestDto.getProductType());
         currentProduct.setName(requestDto.getName());
         currentProduct.setDescription(requestDto.getDescription());
@@ -55,11 +54,18 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void deleteById(Long id) {
+        if (id == null || id <= 0 || productRepository.count() < id) {
+            throw new RuntimeException("Can't find product with id: " + id);
+        }
         productRepository.deleteById(id);
     }
 
+
     @Override
     public List<ProductResponseDto> findProductByName(String name) {
+        if (name == null) {
+            throw new RuntimeException("Name can't be null");
+        }
         return productRepository.findAllByNameContainsIgnoreCase(name)
                 .stream()
                 .map(productMapper::toResponseFromEntity)
@@ -68,6 +74,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public boolean updatePlace(Long productId, String newPlace) {
+        if (newPlace.isEmpty()) {
+            throw new RuntimeException("New place can't be empty");
+        }
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new RuntimeException("Can't find product by id: " + productId));
         product.setPlace(newPlace);
